refactor: split io builtin checker
