+++ src/types/checker/builtins/io/debug.rs
@@ -1,0 +1,28 @@
+use crate::errors::CompileError;
+use crate::parser::ast::Expr;
+use crate::types::{PhpType, TypeEnv};
+
+use super::common::BuiltinResult;
+use super::super::super::Checker;
+
+pub(super) fn check_builtin(
+    checker: &mut Checker,
+    name: &str,
+    args: &[Expr],
+    span: crate::span::Span,
+    env: &TypeEnv,
+) -> BuiltinResult {
+    match name {
+        "var_dump" | "print_r" => {
+            if args.len() != 1 {
+                return Err(CompileError::new(
+                    span,
+                    &format!("{}() takes exactly 1 argument", name),
+                ));
+            }
+            checker.infer_type(&args[0], env)?;
+            Ok(Some(PhpType::Void))
+        }
+        _ => Ok(None),
+    }
+}
