+++ src/types/checker/builtins/io/files.rs
@@ -1,0 +1,176 @@
+use crate::errors::CompileError;
+use crate::parser::ast::Expr;
+use crate::types::{PhpType, TypeEnv};
+
+use super::common::BuiltinResult;
+use super::super::super::Checker;
+
+pub(super) fn check_builtin(
+    checker: &mut Checker,
+    name: &str,
+    args: &[Expr],
+    span: crate::span::Span,
+    env: &TypeEnv,
+) -> BuiltinResult {
+    match name {
+        "file_get_contents" => {
+            if args.len() != 1 {
+                return Err(CompileError::new(
+                    span,
+                    "file_get_contents() takes exactly 1 argument",
+                ));
+            }
+            checker.infer_type(&args[0], env)?;
+            Ok(Some(PhpType::Union(vec![PhpType::Str, PhpType::Bool])))
+        }
+        "file_put_contents" => {
+            if args.len() != 2 {
+                return Err(CompileError::new(
+                    span,
+                    "file_put_contents() takes exactly 2 arguments",
+                ));
+            }
+            for arg in args {
+                checker.infer_type(arg, env)?;
+            }
+            Ok(Some(PhpType::Int))
+        }
+        "file" => {
+            if args.len() != 1 {
+                return Err(CompileError::new(span, "file() takes exactly 1 argument"));
+            }
+            checker.infer_type(&args[0], env)?;
+            Ok(Some(PhpType::Array(Box::new(PhpType::Str))))
+        }
+        "copy" | "rename" => {
+            if args.len() != 2 {
+                return Err(CompileError::new(
+                    span,
+                    &format!("{}() takes exactly 2 arguments", name),
+                ));
+            }
+            for arg in args {
+                checker.infer_type(arg, env)?;
+            }
+            Ok(Some(PhpType::Bool))
+        }
+        "unlink" | "mkdir" | "rmdir" | "chdir" => {
+            if args.len() != 1 {
+                return Err(CompileError::new(
+                    span,
+                    &format!("{}() takes exactly 1 argument", name),
+                ));
+            }
+            checker.infer_type(&args[0], env)?;
+            Ok(Some(PhpType::Bool))
+        }
+        "scandir" | "glob" => {
+            if args.len() != 1 {
+                return Err(CompileError::new(
+                    span,
+                    &format!("{}() takes exactly 1 argument", name),
+                ));
+            }
+            checker.infer_type(&args[0], env)?;
+            Ok(Some(PhpType::Array(Box::new(PhpType::Str))))
+        }
+        "getcwd" => {
+            if !args.is_empty() {
+                return Err(CompileError::new(span, "getcwd() takes no arguments"));
+            }
+            Ok(Some(PhpType::Str))
+        }
+        "tempnam" => {
+            if args.len() != 2 {
+                return Err(CompileError::new(span, "tempnam() takes exactly 2 arguments"));
+            }
+            for arg in args {
+                checker.infer_type(arg, env)?;
+            }
+            Ok(Some(PhpType::Str))
+        }
+        "sys_get_temp_dir" => {
+            if !args.is_empty() {
+                return Err(CompileError::new(
+                    span,
+                    "sys_get_temp_dir() takes no arguments",
+                ));
+            }
+            Ok(Some(PhpType::Str))
+        }
+        "chmod" => {
+            if args.len() != 2 {
+                return Err(CompileError::new(
+                    span,
+                    &format!("{}() takes exactly 2 arguments", name),
+                ));
+            }
+            checker.infer_type(&args[0], env)?;
+            let mode_ty = checker.infer_type(&args[1], env)?;
+            if mode_ty != PhpType::Int {
+                return Err(CompileError::new(args[1].span, "chmod() mode must be int"));
+            }
+            Ok(Some(PhpType::Bool))
+        }
+        "chown" | "chgrp" => {
+            if args.len() != 2 {
+                return Err(CompileError::new(
+                    span,
+                    &format!("{}() takes exactly 2 arguments", name),
+                ));
+            }
+            checker.infer_type(&args[0], env)?;
+            let principal_ty = checker.infer_type(&args[1], env)?;
+            if !matches!(principal_ty, PhpType::Int | PhpType::Str) {
+                return Err(CompileError::new(
+                    args[1].span,
+                    &format!("{}() owner/group must be int or string", name),
+                ));
+            }
+            Ok(Some(PhpType::Bool))
+        }
+        "umask" => {
+            if args.len() > 1 {
+                return Err(CompileError::new(span, "umask() takes 0 or 1 arguments"));
+            }
+            for arg in args {
+                checker.infer_type(arg, env)?;
+            }
+            Ok(Some(PhpType::Int))
+        }
+        "touch" => check_touch(checker, args, span, env).map(Some),
+        _ => Ok(None),
+    }
+}
+
+fn check_touch(
+    checker: &mut Checker,
+    args: &[Expr],
+    span: crate::span::Span,
+    env: &TypeEnv,
+) -> Result<PhpType, CompileError> {
+    if args.is_empty() || args.len() > 3 {
+        return Err(CompileError::new(span, "touch() takes 1, 2, or 3 arguments"));
+    }
+    checker.infer_type(&args[0], env)?;
+    let mut timestamp_types = Vec::new();
+    for arg in args.iter().skip(1) {
+        let ty = checker.infer_type(arg, env)?;
+        if !matches!(ty, PhpType::Int | PhpType::Void) {
+            return Err(CompileError::new(
+                arg.span,
+                "touch() timestamp arguments must be int or null",
+            ));
+        }
+        timestamp_types.push(ty);
+    }
+    if matches!(timestamp_types.first(), Some(PhpType::Void))
+        && matches!(timestamp_types.get(1), Some(ty) if !matches!(ty, PhpType::Void))
+    {
+        return Err(CompileError::new(
+            span,
+            "touch() mtime cannot be null when atime is provided",
+        ));
+    }
+    Ok(PhpType::Bool)
+}
